*Changed VanillaEncoder, now CVT and CCT models use VanillaEncoder.
