+++ vformer/encoder/__init__.py
@@ -1,5 +1,4 @@
 from .cross import CrossEncoder
-from .cvt import CVTEncoderBlock
 from .embedding import *
 from .nn import FeedForward
 from .pyramid import PVTEncoder
