+++ vformer/encoder/vanilla.py
@@ -1,4 +1,5 @@
 import torch.nn as nn
+from timm.models.layers import DropPath
 
 from ..attention import VanillaSelfAttention
 from ..functional import PreNorm
@@ -21,9 +22,23 @@ class VanillaEncoder(nn.Module):
         Dimension of the hidden layer in the feed-forward layer
     p_dropout: float
         Dropout Probability
+    attn_dropout: float
+        Dropout Probability
+    drop_path_rate: float
+        Stochastic drop path rate
     """
 
-    def __init__(self, latent_dim, depth, heads, dim_head, mlp_dim, p_dropout=0.0):
+    def __init__(
+        self,
+        latent_dim,
+        depth,
+        heads,
+        dim_head,
+        mlp_dim,
+        p_dropout=0.0,
+        attn_dropout=0.0,
+        drop_path_rate=0.0,
+    ):
         super().__init__()
         self.encoder = nn.ModuleList([])
         for _ in range(depth):
@@ -36,7 +51,7 @@ class VanillaEncoder(nn.Module):
                                 latent_dim,
                                 heads=heads,
                                 dim_head=dim_head,
-                                p_dropout=p_dropout,
+                                p_dropout=attn_dropout,
                             ),
                         ),
                         PreNorm(
@@ -46,10 +61,15 @@ class VanillaEncoder(nn.Module):
                     ]
                 )
             )
+        self.drop_path = (
+            DropPath(drop_prob=drop_path_rate)
+            if drop_path_rate > 0.0
+            else nn.Identity()
+        )
 
     def forward(self, x):
         for attn, ff in self.encoder:
             x = attn(x) + x
-            x = ff(x) + x
+            x = self.drop_path(ff(x)) + x
 
         return x
